endpoint-client/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
requests==2.28.2
websocket-client==1.5.1
python-dotenv==1.0.0
pymongo==4.3.3
pyjwt==2.6.0
schedule==1.1.0
